refactor: move different targets to different files (#1722)
