+++ cmd/oras/internal/option/interfaces.go
@@ -1,0 +1,35 @@
+/*
+Copyright The ORAS Authors.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package option
+
+import (
+	"oras.land/oras-go/v2"
+	"oras.land/oras-go/v2/content"
+	"oras.land/oras-go/v2/registry"
+)
+
+// ResolvableDeleter a target that resolve and delete
+type ResolvableDeleter interface {
+	content.Resolver
+	content.Deleter
+}
+
+// ReadOnlyGraphTagFinderTarget represents a read-only graph target with tag
+// finder capability.
+type ReadOnlyGraphTagFinderTarget interface {
+	oras.ReadOnlyGraphTarget
+	registry.TagLister
+}
